added sep files on graphs
