+++ Sem. 11/Graphs/Adjacency_list_graph/AdjListGraph.cpp
@@ -1,24 +1,6 @@
 #pragma once
-#include "..\Graph.h"
-#include <list>
-class AdjListGraph : public Graph
-{
-	struct Edge
-	{
-		int end;
-		int weight;
-	};
-	std::vector<std::list<Edge> > adj;
-public:
-	AdjListGraph(int n, bool oriented);
+#include "AdjListGraph.h"
 
-	int addVertex(); 
-
-	void addEdge(int start, int end, int weight = 1);
-	void removeEdge(int start, int end);
-
-	void getSuccessors(int vertex, std::vector<std::pair<int, int>>& vertexAdj) const;
-};
 AdjListGraph::AdjListGraph(int n, bool oriented) : Graph(n, oriented), adj(n)
 {}
 int AdjListGraph::addVertex() //O(1)
