+++ Sem. 11/Graphs/Algorithms/GraphAlgorithms.h
@@ -1,0 +1,7 @@
+#include "..\Graph.h"
+#include <queue>
+#include <stack>
+#include <functional>
+
+void BFS(const Graph& g, int start);
+void DFS(const Graph& g, int start);
