+++ Sem. 11/Graphs/Edge_list_graph/EdgeListGraph.cpp
@@ -1,23 +1,6 @@
-#include "..\Graph.h"
-#include <list>
-class EdgeListGraph : public Graph
-{
-	struct Edge
-	{
-		int start;
-		int end;
-		int weight;
-	};
-	std::list<Edge> edges;
-public:
-	EdgeListGraph(int n, bool oriented);
-	int addVertex(); //return the index of the new vertex
 
-	void addEdge(int start, int end, int weight = 1);
-	void removeEdge(int start, int end);
+#include "EdgeListGraph.h"
 
-	void getSuccessors(int vertex, std::vector<std::pair<int, int>>& vertexAdj) const;
-};
 EdgeListGraph::EdgeListGraph(int n, bool oriented) : Graph(n, oriented)
 {}
