+++ Sem. 11/Graphs/Edge_list_graph/EdgeListGraph.h
@@ -1,0 +1,20 @@
+#include "..\Graph.h"
+#include <list>
+class EdgeListGraph : public Graph
+{
+	struct Edge
+	{
+		int start;
+		int end;
+		int weight;
+	};
+	std::list<Edge> edges;
+public:
+	EdgeListGraph(int n, bool oriented);
+	int addVertex(); //return the index of the new vertex
+
+	void addEdge(int start, int end, int weight = 1);
+	void removeEdge(int start, int end);
+
+	void getSuccessors(int vertex, std::vector<std::pair<int, int>>& vertexAdj) const;
+};
