+++ Sem. 11/Graphs/Graph.cpp
@@ -1,3 +1,4 @@
+#pragma once
 #include "Graph.h"
 
 bool Graph::existsVertex(int vertex) const
